backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
python-dotenv==1.0.1
django-cors-headers==4.6.0
gunicorn==23.0.0
